Adding support for building SIP messages from a stream
Now that all SIP messages, and everything else within siplib, are immutable objects the support for reading a SIP message off a stream of bytes, such as for TCP, had to be re-done. Constructing a SIP message off of a stream is much harder than when you know you have everything that possible would be available for the message right away, which is the case for UDP, so not only was a new stream builder needed but also many of the methods provided in the SipParser had to be made "stream friendly".