Comprehensive rewrite of User Info/Host/Port parsing with robust validation.
Functionality in SipParser.consumeUserInfoHostPort() has been broken out into the SipUserHostInfo class.  This logic has been re-factored and expanded to validate strict conformance to the BNF's for the host portion as specified in RFC 3261.  Supports IPv4 and hostnames, IPv6 support may be added later.  The signature for consumeUserInfoHostPort has been changed to return the new SipUserHostInfo class, which provides access to the port broken out.