Added new ReadOnlyBuffer
There are many cases, specially when we have read something off of a
pcap or the network, where we simply will not expand the original
backing array (which may be a lot of things but commonly a simple
byte-array). In those cases you may want to ensure that you can only
read from this buffer, not write, and if so, this new ReadOnlyBuffer is
for you. Since you cannot change the underlying buffer, it is cheap to
do slice operations etc and clone will still also make use of the same
backing array but with its own reader index.