Added support for extracting the seq no
Slightly different API for GTPv1 vs GTPv2 because in v1, seq no is
optional, whereas in GTPv2 it isn't. The seq no in v1 is also only 2
bytes but in v2 it is three bytes.

Hence, instead of having those methods on the top-level GtpHeader
interfaces, they are on the versioned ones...