Added basic support for Information Elements
Each GTP message consists of a header, followed by 1 or more Information
Elements (IE). Depending on the version of GTP (1 or 2), the type of IE
will differ and even within GTPv1, there are different IEs.

This commit adds basic support for the structure of those IEs and one
basic version of the GTPv2 TLIV - Type, Length, Instance, Value.