Java 11 now minimum version, upgrading dependencies.
As there has been non-backwards compatible changes, especially after
Java 9, in regards to how Mockito can, and is allowed, to mock things, I
was forced to upgrade Mockito, which also forced me to move to Java 11
as the minimum version. If not, it seems that people would be stuck on
using Java 8 for this library with very old versions of e.g. mockito.

Java 8, althoug still supported, is way old so let's move forward. Let's
face it, Java 11 is also pretty old at this point.